Split body.Task into a separate module
